[dy] Reorganize
