+++ mage_integrations/mage_integrations/sources/chargebee/__init__.py
@@ -1,24 +1,16 @@
 from mage_integrations.sources.base import Source, main
 from mage_integrations.sources.catalog import Catalog
 from mage_integrations.sources.chargebee.client import ChargebeeClient
-from mage_integrations.sources.chargebee.client.tap_chargebee.streams import (
+from mage_integrations.sources.chargebee.state import save_state
+from mage_integrations.sources.chargebee.streams import (
     ITEM_MODEL_AVAILABLE_STREAMS,
     PLAN_MODEL_AVAILABLE_STREAMS,
 )
-from mage_integrations.sources.chargebee.client.tap_chargebee.streams.base import is_selected
-from mage_integrations.sources.messages import write_state
+from mage_integrations.sources.chargebee.streams.base import is_selected
 from typing import List
 import singer
 
 LOGGER = singer.get_logger()
-
-def save_state(state):
-    if not state:
-        return
-
-    LOGGER.info('Updating state.')
-
-    write_state(state)
 
 class Runner:
