+++ mage_integrations/mage_integrations/sources/chargebee/streams/subscriptions.py
@@ -1,10 +1,9 @@
-from mage_integrations.sources.chargebee.client.tap_chargebee.streams.base \
-    import BaseChargebeeStream
+from mage_integrations.sources.chargebee.streams.base import BaseChargebeeStream
 
 
-class CouponsStream(BaseChargebeeStream):
-    TABLE = 'coupons'
-    ENTITY = 'coupon'
+class SubscriptionsStream(BaseChargebeeStream):
+    TABLE = 'subscriptions'
+    ENTITY = 'subscription'
     REPLICATION_METHOD = 'INCREMENTAL'
     REPLICATION_KEY = 'updated_at'
     KEY_PROPERTIES = ['id']
@@ -13,13 +12,13 @@ class CouponsStream(BaseChargebeeStream):
     VALID_REPLICATION_KEYS = ['updated_at']
     INCLUSION = 'available'
     API_METHOD = 'GET'
-    SCHEMA = 'plan_model/coupons'
-    SORT_BY = 'created_at'
+    SCHEMA = 'plan_model/subscriptions'
+    SORT_BY = 'updated_at'
 
     def __init__(self, config, state, catalog, client):
         BaseChargebeeStream.__init__(self, config, state, catalog, client)
         if self.config['item_model']:
-            self.SCHEMA = 'item_model/coupons'
+            self.SCHEMA = 'item_model/subscriptions'
 
     def get_url(self):
-        return 'https://{}.chargebee.com/api/v2/coupons'.format(self.config.get('site'))
+        return 'https://{}.chargebee.com/api/v2/subscriptions'.format(self.config.get('site'))
