+++ mage_integrations/mage_integrations/sources/chargebee/templates/config.json
@@ -2,7 +2,7 @@
   "start_date": "2022-01-01T00:00:00Z",
   "api_key": "",
   "site": "",
-  "include_deleted": "False",
+  "include_deleted": "True",
   "request_timeout": 300
 }
